Fix issue with huge times
When a match list is loaded, the unrun matches were being sorted first since they defaulted to epoc 0 time. This caused everything to have massive timestamps until all the matches had been run.

The change updates sorting time to behave in a more reasonable fashion and puts N/A in for un-run matches.